@@ -1,5 +1,3 @@
 module go.felesatra.moe/pwnck
 
-require go.felesatra.moe/go2/errors v1.0.0
-
-go 1.13
+go 1.22.4
